Fix: 토큰 생성기 수정
