+++ tokens/design_system_semantic.json
@@ -15,7 +15,7 @@
     "backgroud": {
       "normal": {
         "$type": "color",
-        "$value": "{color.common.white}"
+        "$value": "{color.white}"
       },
       "alternative": {
         "$type": "color",
@@ -39,7 +39,7 @@
       },
       "strong": {
         "$type": "color",
-        "$value": "{color.common.black}"
+        "$value": "{color.black}"
       },
       "neutral": {
         "$type": "color",
@@ -84,6 +84,80 @@
     }
   },
   "typography": {
+    "heading18": {
+      "extraBold": {
+        "family": {
+          "$type": "text",
+          "$value": "{font.family.pretendard}"
+        },
+        "weight": {
+          "$type": "number",
+          "$value": "{font.weight.extraBold}"
+        },
+        "lineHeight": {
+          "$type": "number",
+          "$value": "{font.lineHeight.26}"
+        },
+        "size": {
+          "$type": "number",
+          "$value": "{font.size.18}"
+        }
+      },
+      "bold": {
+        "family": {
+          "$type": "text",
+          "$value": "{font.family.pretendard}"
+        },
+        "weight": {
+          "$type": "number",
+          "$value": "{font.weight.bold}"
+        },
+        "lineHeight": {
+          "$type": "number",
+          "$value": "{font.lineHeight.26}"
+        },
+        "size": {
+          "$type": "number",
+          "$value": "{font.size.18}"
+        }
+      },
+      "medium": {
+        "family": {
+          "$type": "text",
+          "$value": "{font.family.pretendard}"
+        },
+        "weight": {
+          "$type": "number",
+          "$value": "{font.weight.medium}"
+        },
+        "lineHeight": {
+          "$type": "number",
+          "$value": "{font.lineHeight.26}"
+        },
+        "size": {
+          "$type": "number",
+          "$value": "{font.size.18}"
+        }
+      },
+      "regular": {
+        "family": {
+          "$type": "text",
+          "$value": "{font.family.pretendard}"
+        },
+        "weight": {
+          "$type": "number",
+          "$value": "{font.weight.regular}"
+        },
+        "lineHeight": {
+          "$type": "number",
+          "$value": "{font.lineHeight.26}"
+        },
+        "size": {
+          "$type": "number",
+          "$value": "{font.size.18}"
+        }
+      }
+    },
     "title32": {
       "medium": {
         "family": {
@@ -249,80 +323,6 @@
         "size": {
           "$type": "number",
           "$value": "{font.size.20}"
-        }
-      }
-    },
-    "heading18": {
-      "extraBold": {
-        "family": {
-          "$type": "text",
-          "$value": "{font.family.pretendard}"
-        },
-        "weight": {
-          "$type": "number",
-          "$value": "{font.weight.extraBold}"
-        },
-        "lineHeight": {
-          "$type": "number",
-          "$value": "{font.lineHeight.26}"
-        },
-        "size": {
-          "$type": "number",
-          "$value": "{font.size.18}"
-        }
-      },
-      "bold": {
-        "family": {
-          "$type": "text",
-          "$value": "{font.family.pretendard}"
-        },
-        "weight": {
-          "$type": "number",
-          "$value": "{font.weight.bold}"
-        },
-        "lineHeight": {
-          "$type": "number",
-          "$value": "{font.lineHeight.26}"
-        },
-        "size": {
-          "$type": "number",
-          "$value": "{font.size.18}"
-        }
-      },
-      "medium": {
-        "family": {
-          "$type": "text",
-          "$value": "{font.family.pretendard}"
-        },
-        "weight": {
-          "$type": "number",
-          "$value": "{font.weight.medium}"
-        },
-        "lineHeight": {
-          "$type": "number",
-          "$value": "{font.lineHeight.26}"
-        },
-        "size": {
-          "$type": "number",
-          "$value": "{font.size.18}"
-        }
-      },
-      "regular": {
-        "family": {
-          "$type": "text",
-          "$value": "{font.family.pretendard}"
-        },
-        "weight": {
-          "$type": "number",
-          "$value": "{font.weight.regular}"
-        },
-        "lineHeight": {
-          "$type": "number",
-          "$value": "{font.lineHeight.26}"
-        },
-        "size": {
-          "$type": "number",
-          "$value": "{font.size.18}"
         }
       }
     },
